refactor(InkVideoPlayer): 抽出 inkVideoPlayer/ 子模块（纯逻辑/HUD/浮层/投屏 hook）
- playback.ts: 格式判定、倍速/手势常量、旋转与视图/手势状态模型（无 React）
- GestureHudOverlay.tsx / PlayerBatteryIcon.tsx: 手势 HUD 与电量图标展示组件
- MediaResourceOverlay.tsx / CastOverlay.tsx: 嗅探资源抽屉与 DLNA 投屏浮层
- useCastControls.ts: DLNA 设备发现/连接/遥控/结束的状态 hook
- 主文件保留 wrapper + core 状态机（2504 → 1751 行），全屏/手势/嗅探路径不动
- 修复播放失败文案乱码（瑙嗛婧愭殏鏃舵棤娉曟挱鏀? → 视频源暂时无法播放）
- 同步 dlna-cast-lifecycle / origin-player-live-sniff 静态测试读取路径（断言不弱化）
